requirements.txt
aiohttp
click
python-dateutil
pyyaml